+++ fla/models/hgrn2/modeling_hgrn2.py
@@ -223,7 +223,7 @@ class HGRN2Model(HGRN2PreTrainedModel):
         all_attns = () if output_attentions else None
 
         if self.config.use_lower_bound:
-            lower_bounds = self.lower_bounds.softmax(0)
+            lower_bounds = self.lower_bounds.softmax(0, dtype=torch.float)
             lower_bounds = lower_bounds.cumsum(0) - lower_bounds[0]
         for i, layer in enumerate(self.layers):
             if output_hidden_states:
